+++ full-backup.sh
@@ -3,7 +3,7 @@
 # requirements
 # ~/log, ~/backups, ~/path/to/example.com/public
 
-version=6.0.3
+version=6.1.1
 
 # this script is basically
 #   files-backup-without-uploads.sh script + part of db-backup.sh script
@@ -26,7 +26,12 @@ PASSPHRASE=
 # if you have a different pattern, such as ~/app/example.com, please change the following to fit the server environment!
 SITES_PATH=${HOME}/sites
 
+# To debug, use any value for "debug", otherwise please leave it empty
+debug=
+
 #-------- Do NOT Edit Below This Line --------#
+
+[ "$debug" ] && set -x
 
 log_file=${HOME}/log/backups.log
 exec > >(tee -a "${log_file}")
@@ -43,8 +48,11 @@ DOMAIN=
 PUBLIC_DIR=public
 
 # get environment variables, if exists
+# .envrc is in the following format
+# export VARIABLE=value
 [ -f "$HOME/.envrc" ] && source ~/.envrc
-[ -f "$HOME/.env" ] && source ~/.env
+# uncomment the following, if you use .env with the format "VARIABLE=value" (without export)
+# if [ -f "$HOME/.env" ]; then; set -a; source ~/.env; set +a; fi
 
 # printf 'Usage: %s [-b|--bucket <name>] [-k|--keepfor <days>] [-e|--email <email-address>] [-p|--path <WP path>] [-v|--version] [-h|--help] example.com\n' "$0"
 print_help() {
@@ -204,8 +212,17 @@ for i in "${!EXC_PATH[@]}" ; do
     # remember the trailing space; we'll use it later
 done
 
+if [ "$debug" ]; then
+    echo "exclude_base_path: $exclude_base_path"
+    printf "EXC_PATH: %s\n" "${EXC_PATH[@]}"
+    echo "EXCLUDES: $EXCLUDES"
+
+    # exit
+fi
+
 #------------- from db-script.sh --------------#
 # take actual DB backup
+# 2>/dev/null to suppress any warnings / errors
 wp --path="${WP_PATH}" transient delete --all
 if ! wp --path="${WP_PATH}" db export --no-tablespaces=true --add-drop-table "$db_dump"; then
     msg="$script_name - [Error] Something went wrong while taking DB dump!"
@@ -225,12 +242,12 @@ if [ "$PASSPHRASE" ]; then
     LATEST_FULL_BACKUP_FILE_NAME=${LATEST_FULL_BACKUP_FILE_NAME}.gpg
     # using symmetric encryption
     # option --batch to avoid passphrase prompt
-    tar hcz "${EXCLUDES}" -C "${SITES_PATH}" "${dir_to_backup}" | gpg --symmetric --passphrase "$PASSPHRASE" --batch -o "$FULL_BACKUP_FILE_NAME"
+    tar hcz -C "${SITES_PATH}" ${EXCLUDES} "${dir_to_backup}" | gpg --symmetric --passphrase "$PASSPHRASE" --batch -o "$FULL_BACKUP_FILE_NAME"
 else
     echo "[Warn] No passphrase provided for encryption!"
     echo "[Warn] If you are from Europe, please check GDPR compliance."
     # tar hczf ${FULL_BACKUP_FILE_NAME} --warning=no-file-changed ${EXCLUDES} -C ${SITES_PATH} ${dir_to_backup} > /dev/null
-    tar hczf "${FULL_BACKUP_FILE_NAME}" "${EXCLUDES}" -C "${SITES_PATH}" "${dir_to_backup}" > /dev/null
+    tar hczf "${FULL_BACKUP_FILE_NAME}" ${EXCLUDES} -C "${SITES_PATH}" "${dir_to_backup}" > /dev/null
 fi
 if [ "$?" = "0" ]; then
     printf "\nBackup is successfully taken locally.\n\n"